Factor TarjanSCCCalculator out of QI checker.
