+++ src/herder/QuorumIntersectionCheckerImpl.cpp
@@ -53,82 +53,6 @@ QBitSet::getSuccessors(BitSet const& nodes, QGraph const& inner)
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-// Implementation of TarjanSCCCalculator
-////////////////////////////////////////////////////////////////////////////////
-//
-// This is a completely stock implementation of Tarjan's algorithm for
-// calculating strongly connected components. Like "read off of wikipedia"
-// stock. Go have a look!
-//
-// https://en.wikipedia.org/wiki/Tarjan%27s_strongly_connected_components_algorithm
-
-TarjanSCCCalculator::TarjanSCCCalculator(QGraph const& graph) : mGraph(graph)
-{
-}
-
-void
-TarjanSCCCalculator::calculateSCCs()
-{
-    mNodes.clear();
-    mStack.clear();
-    mIndex = 0;
-    mSCCs.clear();
-    for (size_t i = 0; i < mGraph.size(); ++i)
-    {
-        mNodes.emplace_back(SCCNode{});
-    }
-    for (size_t i = 0; i < mGraph.size(); ++i)
-    {
-        if (mNodes.at(i).mIndex == -1)
-        {
-            scc(i);
-        }
-    }
-}
-
-void
-TarjanSCCCalculator::scc(size_t i)
-{
-    auto& v = mNodes.at(i);
-    v.mIndex = mIndex;
-    v.mLowLink = mIndex;
-    mIndex++;
-    mStack.push_back(i);
-    v.mOnStack = true;
-
-    BitSet const& succ = mGraph.at(i).mAllSuccessors;
-    for (size_t j = 0; succ.nextSet(j); ++j)
-    {
-        CLOG_TRACE(SCP, "edge: {} -> {}", i, j);
-        SCCNode& w = mNodes.at(j);
-        if (w.mIndex == -1)
-        {
-            scc(j);
-            v.mLowLink = std::min(v.mLowLink, w.mLowLink);
-        }
-        else if (w.mOnStack)
-        {
-            v.mLowLink = std::min(v.mLowLink, w.mIndex);
-        }
-    }
-
-    if (v.mLowLink == v.mIndex)
-    {
-        BitSet newScc;
-        newScc.set(i);
-        size_t j = 0;
-        do
-        {
-            j = mStack.back();
-            newScc.set(j);
-            mStack.pop_back();
-            mNodes.at(j).mOnStack = false;
-        } while (j != i);
-        mSCCs.push_back(newScc);
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////
 // Implementation of MinQuorumEnumerator
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -347,11 +271,13 @@ QuorumIntersectionCheckerImpl::QuorumIntersectionCheckerImpl(
     : mCfg(cfg)
     , mLogTrace(Logging::logTrace("SCP"))
     , mQuiet(quiet)
-    , mTSC(mGraph)
+    , mTSC()
     , mInterruptFlag(interruptFlag)
     , mCachedQuorums(MAX_CACHED_QUORUMS_SIZE)
 {
     buildGraph(qmap);
+    // Awkwardly, the graph size is zero when we initialize mTSC. Update it
+    // here.
     buildSCCs();
 }
 
@@ -694,7 +620,12 @@ QuorumIntersectionCheckerImpl::buildGraph(QuorumTracker::QuorumMap const& qmap)
 void
 QuorumIntersectionCheckerImpl::buildSCCs()
 {
-    mTSC.calculateSCCs();
+    mTSC.calculateSCCs(mGraph.size(), [this](size_t i) -> BitSet const& {
+        // NB: this closure must be written with the explicit const&
+        // returning type signature, otherwise it infers wrong and
+        // winds up returning a dangling reference at its site of use.
+        return this->mGraph.at(i).mAllSuccessors;
+    });
     mStats.mNumSCCs = mTSC.mSCCs.size();
 }
